Deploy Permissions API Integration v1
* Consume utilities from test

* Setup test workflow

* Import prisma client from the common repo

* Delete prisma from webapp

* Fix test page import

* Source prisma schema from node_modules

* Always generate prisma schema

* delete prisma folder

* Create a symbolic link to the source prisma directory

* Update symlink approach

* Fix type errors

* Update typing source

* Fix prisma client import

* Implement delegated call

* Install the latest version of charmverse core

* Update CI

* Add devdependencies to charmverse install

* trigger CI

* Update deps

* Update build steps

* Add prisma to ignored dependencies

* Add prisma to gitignore

* Update copy command

* Add data about current step

* Add debug step

* Fix typo

* Split actions

* Debug prisma

* Clean the directory

* Update the build process

* Trigger CI

* Trigger CI

* Add charmverse/core to esm modules

* Add permissions API Url to the deploy

* Fix max workers setting

* Generate .env.test.local inside install step

* Update package-lock

* Add router implementation for getting client

* Trigger CI with new permissions endpoint

* Fix imports

* Consume prisma client from charmverse common

* Fix build

* Tweak build steps

* Update imports

* Delete jest mock

* Use updated branch

* Consume the permissions API inside all forum endpoints

* Consume computePostPermissions routing everywhere

* WIP permissions

* Migrate usage of apps

* Implement simple computePostPermissions

* update app

* Upgrade jest and get the app working

* Fix all typescript errors

* Trigger CI

* Trigger CI

* Update to use notion of listing permissioned categories

* Tweak prisma

* Import permissions API

* Add dependency

* Setup directory

* Remove action step

* Fix syntax of action

* Fix syntax of action

* Delete dir prop

* Reorder build

* Reorder build

* FIx build

* FIx build

* FIx build

* FIx build

* FIx build

* FIx build

* Trigger CI

* Fix the app

* Fix the app

* use SSH key

* fix build

* Update build and caching steps

* Fix the github action

* Fix the github action

* Fix the github action

* Fix the github action

* Fix the github action

* Fix the github action

* Fix the github action

* Add port to the permissions API startup

* Fix command

* Debug folder

* Restore permisisons API in the integration step

* Include SSH key on second step

* Drop request operations

* Delete prisma

* Add module name mapper, and commented webpack experiments for fixing npm link

* Save changes

* Replace prisma client import

* Update package-lock

* Update package lock

* Fix endpoints for tests to work

* Fix scripts

* Update dependency version, include core in the cache, link to post install

* Fix jest test teardown

* Get builder from core

* Stable app version

* Fix the types for build

* Config playwright

* Lock working version

* Add extra TSConfig

* Stable prod build

* Commit stable build approach

* WIP debug config

* Fix path mapping

* Fix dependency version

* Update deps

* Bump version

* Update app

* Readd @charmverse/core as an esm module

* Fix broken unit tests

* Increase test timeout

* Increase timeout and remove code

* Fix broken get post endpoint, add missing endpoint tests

* Fix browser tests

* Switch core to main branch

* Build permissions API after main app so we can spot build errors sooner

* Tweak caching for permissions API andtemporarily re-hoist

* Use main branch for building permissions API

* Restablish build steps, ignore permissions API tests

* Import errors from the core lib

* Implement switching logic between public and paid spaces

* Drop duplicate forum interfaces and import from core

* Bump package version

* Update application name

* Reset the github action value

* Fix types

* Pull the trigger

* Fix docker

* Bump package version

* Push latest ref

* Delete uneeded file

* Update deploy script to run tests this time and deploy

* Trigger CI

* Redeploy

* Bump core version

* Deploy Permissions API Integration v1

---------

+++ pages/api/forums/posts/[postId]/index.ts
@@ -1,18 +1,20 @@
-import type { Post } from '@prisma/client';
+import { prisma } from '@charmverse/core';
+import type { Post } from '@charmverse/core/dist/prisma';
 import type { NextApiRequest, NextApiResponse } from 'next';
 import nc from 'next-connect';
 
-import { prisma } from 'db';
 import { deleteForumPost } from 'lib/forums/posts/deleteForumPost';
 import { getForumPost } from 'lib/forums/posts/getForumPost';
 import type { UpdateForumPostInput } from 'lib/forums/posts/updateForumPost';
 import { updateForumPost } from 'lib/forums/posts/updateForumPost';
 import { ActionNotPermittedError, onError, onNoMatch, requireUser } from 'lib/middleware';
-import { requestOperations } from 'lib/permissions/requestOperations';
+import { getPermissionsClient } from 'lib/permissions/api';
 import { withSessionRoute } from 'lib/session/withSession';
+import { isUUID } from 'lib/utilities/strings';
 import { WebhookEventNames } from 'lib/webhookPublisher/interfaces';
 import { publishPostEvent } from 'lib/webhookPublisher/publishEvent';
 import { relay } from 'lib/websockets/relay';
+import computePostPermissions from 'pages/api/permissions/forum/compute-post-permissions';
 
 const handler = nc<NextApiRequest, NextApiResponse>({ onError, onNoMatch });
 
@@ -22,12 +24,19 @@ async function updateForumPostController(req: NextApiRequest, res: NextApiRespon
   const { postId } = req.query as any as { postId: string };
   const userId = req.session.user.id;
 
-  await requestOperations({
-    resourceType: 'post',
-    operations: ['edit_post'],
+  const permissionsClient = await getPermissionsClient({
+    resourceId: postId,
+    resourceIdType: 'post'
+  });
+
+  const permissions = await permissionsClient.forum.computePostPermissions({
     resourceId: postId,
     userId
   });
+
+  if (!permissions.edit_post) {
+    throw new ActionNotPermittedError(`You cannot edit this post`);
+  }
 
   const existingPost = await prisma.post.findUniqueOrThrow({
     where: {
@@ -47,12 +56,14 @@ async function updateForumPostController(req: NextApiRequest, res: NextApiRespon
 
   // Don't allow an update to a new category ID if the user doesn't have permission to create posts in that category
   if (typeof newCategoryId === 'string' && existingPost.categoryId !== newCategoryId) {
-    await requestOperations({
-      resourceType: 'post_category',
-      operations: ['create_post'],
+    const categoryPermissions = await permissionsClient.forum.computePostCategoryPermissions({
       resourceId: newCategoryId,
       userId
     });
+
+    if (!categoryPermissions.create_post) {
+      throw new ActionNotPermittedError(`You cannot change the post category to this new category`);
+    }
   }
 
   const post = await updateForumPost(postId, req.body);
@@ -98,12 +109,16 @@ async function deleteForumPostController(req: NextApiRequest, res: NextApiRespon
   const { postId } = req.query as any as { postId: string };
   const userId = req.session.user.id;
 
-  await requestOperations({
-    resourceType: 'post',
-    resourceId: postId,
-    operations: ['delete_post'],
-    userId
-  });
+  const permissions = await getPermissionsClient({ resourceId: postId, resourceIdType: 'post' }).then(({ forum }) =>
+    forum.computePostPermissions({
+      resourceId: postId,
+      userId
+    })
+  );
+
+  if (!permissions.delete_post) {
+    throw new ActionNotPermittedError(`You cannot delete this post`);
+  }
 
   const post = await deleteForumPost(postId);
 
@@ -129,13 +144,17 @@ async function getForumPostController(req: NextApiRequest, res: NextApiResponse<
 
   const post = await getForumPost({ userId, postId, spaceDomain });
 
-  await requestOperations({
-    resourceType: 'post',
-    resourceId: post.id,
-    operations: ['view_post'],
-    userId,
-    customError: new ActionNotPermittedError(`You do not have access to this post`)
-  });
+  const permissions = await getPermissionsClient({ resourceId: post.id, resourceIdType: 'post' }).then(({ forum }) =>
+    forum.computePostPermissions({
+      resourceId: post.id,
+      userId
+    })
+  );
+
+  if (!permissions.view_post) {
+    throw new ActionNotPermittedError(`You do not have access to this post`);
+  }
+
   res.status(200).json(post);
 }
